CARDS-1337: Enable patient authentication Renamed java package from `tokenguest` to `token`
