+++ modules/token-authentication/src/main/java/io/uhndata/cards/auth/token/TokenManager.java
@@ -15,7 +15,7 @@
  * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package io.uhndata.cards.auth.tokenguest;
+package io.uhndata.cards.auth.token;
 
 import java.util.Calendar;
 import java.util.Map;
