+++ modules/token-authentication/src/main/java/io/uhndata/cards/auth/token/impl/CardsTokenImpl.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.uhndata.cards.auth.tokenguest.impl;
+package io.uhndata.cards.auth.token.impl;
 
 import java.util.Arrays;
 import java.util.Calendar;
