+++ modules/token-authentication/src/main/java/io/uhndata/cards/auth/token/impl/oak/CardsTokenConfigurationImpl.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.uhndata.cards.auth.tokenguest.impl.oak;
+package io.uhndata.cards.auth.token.impl.oak;
 
 import org.apache.jackrabbit.oak.api.Root;
 import org.apache.jackrabbit.oak.spi.security.ConfigurationBase;
